+++ test/unit/kaocha/report_test.clj
@@ -327,8 +327,4 @@
 (comment
   (do
     (require 'kaocha.repl)
-    (kaocha.repl/run)
-
-    )
-
-  )
+    (kaocha.repl/run)))