+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Warn when running Kaocha without a configuration file. This is fine for
     experimenting, but for long-term use, we recommend creating a configuration
     file to avoid changes in behavior between releases.
+- Fix exception when running Kaocha on Windows with the built-in notification
+    plugin enabled.
 
 ## Changed