+++ CHANGELOG.md
@@ -14,6 +14,8 @@
     experimenting, but for long-term use, we recommend creating a configuration
     file to avoid changes in behavior between releases.
 - Provide a warning when no tests are found.
+- Fix exception when running Kaocha on Windows with the built-in notification
+    plugin enabled.
 
 ## Changed