+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.0.829 (2021-03-08 / a88ebda)
 
 ## Added